Update GetCams_Main.sh
If main camera was on /dev/video2, the camera would fail since it was not found. Instead, check if ANY /dev/video* is assigned, then get the results of the ls command and assign that to the camera's logical address on the system. This accounts for any /dev/video* address, but may also crash if you have more than one USB video device.